pyproject.toml
[project]
name = "oneclicksetup"
version = "0.1.0"
description = "A one-click environment setup CLI."
authors = [
  { name="Your Name", email="you@example.com" }
]
readme = "README.md"
keywords = ["cli", "devops", "setup"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  "python-dotenv==0.21.0",  # Pin the version for reproducibility
]
requires-python = ">=3.6"

[project.scripts]
mysetup = "oneclicksetup.cli:main"
